BigData-Deployment-Manager/requirements.txt
# Big Data Deployment Manager Dependencies

# PySpark for big data processing
pyspark==3.5.0

# Flask API framework
Flask==3.0.0
flask-cors==4.0.0

# Database drivers
psycopg2-binary==2.9.9         # PostgreSQL driver
pymysql==1.1.0                 # MySQL driver (optional)

# Configuration management
PyYAML==6.0.1
python-dotenv==1.0.0

# Testing frameworks
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Logging and monitoring
colorlog==6.8.0

# Optional: for data validation
pydantic==2.5.0

# Optional: for caching (if needed)
redis==5.0.1

# Development dependencies
black==23.12.0                 # Code formatting
flake8==6.1.0                  # Linting
mypy==1.7.1                    # Type checking
